Added test fixture proving that bug exists also for parameters
